+++ datar/core/backends/__init__.py
@@ -1,3 +1,0 @@
-from ..options import get_option
-
-BACKEND = get_option("backend")