+++ datar/core/backends/pandas/core/generic.py
@@ -1,1 +1,0 @@
-from pandas.core.generic import NDFrame