+++ datar/core/exceptions.py
@@ -1,5 +1,0 @@
-"""Exceptions for datar"""
-
-
-class NameNonUniqueError(Exception):
-    """When check_unique fails"""