+++ datar/core/options.py
@@ -1,63 +1,31 @@
 """Provide options"""
+from __future__ import annotations
 
-from typing import Any, Generator, Mapping, Union, Callable
+from typing import Any, Generator, Mapping
 from contextlib import contextmanager
-from pathlib import Path
 
 from diot import Diot
 from simpleconf import Config
+
+from .defaults import OPTION_FILE_CWD, OPTION_FILE_HOME
 
 _key_transform = lambda key: key.replace("_", ".")
 _dict_transform_back = lambda dic: {
     key.replace(".", "_"): val for key, val in dic.items()
 }
 
-
-def enable_pdtypes_callback(enable: bool) -> None:  # pragma: no cover
-    from .utils import logger
-
-    try:
-        import pdtypes
-    except ImportError:
-        pdtypes = None
-
-    if enable and pdtypes is None:
-        logger.warning(
-            "Package `pdtypes` not installed for `options(enable_pdtypes=True)`"
-        )
-    elif not enable and pdtypes is not None:
-        pdtypes.unpatch()
-
-
-OPTION_CALLBACKS = Diot(
-    # allow 'a.b' to access 'a_b'
-    diot_transform=_key_transform,
-    enable_pdtypes=enable_pdtypes_callback,
-)
-
-
-OPTION_FILE_HOME = Path("~/.datar.toml").expanduser()
-OPTION_FILE_CWD = Path("./.datar.toml").resolve()
-OPTIONS = Diot(
-    # Whether use 0-based numbers when index is involved, acts similar like R
-    dplyr_summarise_inform=True,
-    # What to do when there are conflicts importing names
-    # - `warn`: show warnings
-    # - `silent`: ignore the conflicts
-    # - `underscore_suffixed`: add suffix `_` to the conflicting names
-    #   (and don't do any warnings)
-    import_names_conflict="warn",
-    # Enable pdtypes
-    enable_pdtypes=True,
-    # add_option=True,
-    # allow 'a.b' to access 'a_b'
-    # diot_transform=_key_transform,
-    # The backend for datar
-    backend="pandas",  # or "modin"
-)
 OPTIONS = Diot(
     Config.load(
-        OPTIONS,
+        {
+            # What to do when there are conflicts importing names
+            # - `warn`: show warnings
+            # - `silent`: ignore the conflicts
+            # - `underscore_suffixed`: add suffix `_` to the conflicting names
+            #   (and don't do any warnings)
+            "import_names_conflict": "warn",
+            # Disable some installed backends
+            "backends": [],
+        },
         OPTION_FILE_HOME,
         OPTION_FILE_CWD,
         ignore_nonexist=True,
@@ -66,14 +34,8 @@ OPTIONS = Diot(
 )
 
 
-def apply_init_callbacks():
-    """Apply the callbacks when options are initialized"""
-    for key in OPTION_CALLBACKS:
-        OPTION_CALLBACKS[key](OPTIONS[key])
-
-
 def options(
-    *args: Union[str, Mapping[str, Any]],
+    *args: str | Mapping[str, Any],
     _return: bool = None,
     **kwargs: Any,
 ) -> Mapping[str, Any]:
@@ -89,6 +51,7 @@ def options(
         The options before updating if `_return` is `True`.
     """
     if not args and not kwargs and (_return is None or _return is True):
+        # Make sure the options won't be changed
         return OPTIONS.copy()
 
     names = [arg.replace(".", "_") for arg in args if isinstance(arg, str)]
@@ -117,9 +80,6 @@ def options(
         if oldval == val:
             continue
         OPTIONS[key] = val
-        callback = OPTION_CALLBACKS.get(key)
-        if callable(callback):  # pragma: no cover, already applied
-            callback(val)
 
     return out
 
@@ -137,8 +97,8 @@ def options_context(**kwargs: Any) -> Generator:
 
 
 def get_option(x: str, default: Any = None) -> Any:
-    """Get the current value set for option ‘x’,
-    or ‘default’ (which defaults to ‘NULL’) if the option is unset.
+    """Get the current value set for option `x`,
+    or `default` (which defaults to `NULL`) if the option is unset.
 
     Args:
         x: The name of the option
@@ -147,9 +107,11 @@ def get_option(x: str, default: Any = None) -> Any:
     return OPTIONS.get(x, default)
 
 
-def add_option(x: str, default: Any = None, callback: Callable = None) -> None:
-    """Add an option"""
-    OPTIONS[x] = default
-    if callback:
-        OPTION_CALLBACKS[x] = callback
-        callback(default)
+def add_option(x: str, default: Any = None) -> None:
+    """Add an option
+
+    Args:
+        x: The name of the option
+        default: The default value if `x` is unset
+    """
+    OPTIONS.setdefault(x, default)