+++ datar/forcats.py
@@ -1,0 +1,5 @@
+from .core.plugin import plugin as _plugin
+from .apis.forcats import *
+
+_additional_imports = _plugin.hooks.forcats_api()
+locals().update(_additional_imports)